init: add modprobe
We added the modprobe function to init before but it was dropped in
oreo-x86. Originally I planned to replace it by toybox's modprobe.
However, it doesn't work as expected. For example, the audio driver
is not loaded correctly in some Skylake devices.

To fix that, add back the modprobe function to init.

Signed-off-by: Shanmin Zhang <zhangshanmin@mail.tsinghua.edu.cn>